E-Commerce na Azure/requirements.txt
streamlit
azure-storage-blob
pymssql
python-dotenv